Move splash front page to own file, only use hooks
