+++ frontend/index.html
@@ -5,6 +5,7 @@
         <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap" />
     </head>
     <body>
+        <div id="root"></div>
         <script src="main.tsx"></script>
     </body>
 </html
