+++ frontend/package.json
@@ -25,7 +25,6 @@
     "random-js": "^2.1.0",
     "react": "^16.13.0",
     "react-dom": "^16.13.0",
-    "reactive-lens": "^0.5.0",
     "vega": "^5.10.0",
     "vega-lite": "^4.7.0"
   }
